# praticas/praticas09/swagger.yaml
openapi: 3.0.0
info:
  title: API de Produtos
  description: Documentação da API de produtos da prática 09.
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Ambiente de desenvolvimento

tags:
  - name: produtos
    description: Operações relacionadas a produtos

paths:
  /produtos:
    get:
      tags:
        - produtos
      summary: Lista todos os produtos
      responses:
        '200':
          description: Uma lista de produtos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArrayOfProdutos'

    post:
      tags:
        - produtos
      summary: Adiciona um novo produto
      requestBody:
        description: Um JSON com dados do novo produto
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Produto'
      responses:
        '201':
          description: Produto criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '422':
          description: Nome e preço são obrigatórios

  /produtos/{produtoId}:
    get:
      tags:
        - produtos
      summary: Busca um produto pelo ID
      parameters:
        - in: path
          name: produtoId
          required: true
          description: Id do produto a ser buscado
          schema:
            type: integer
      responses:
        '200':
          description: Um JSON com os dados do produto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '404':
          description: Produto não encontrado

    put:
      tags:
        - produtos
      summary: Atualiza um produto pelo ID
      parameters:
        - in: path
          name: produtoId
          required: true
          description: Id do produto a ser buscado
          schema:
            type: integer
      requestBody:
        description: Um JSON com dados do produto a atualizar
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Produto'
      responses:
        '200':
          description: Um JSON com os dados do produto atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '404':
          description: Produto não encontrado
        '422':
          description: Nome e preço são obrigatórios

    delete:
      tags:
        - produtos
      summary: Remove um produto pelo ID
      parameters:
        - in: path
          name: produtoId
          required: true
          description: Id do produto a ser buscado
          schema:
            type: integer
      responses:
        '204':
          description: Produto removido com sucesso
        '404':
          description: Produto não encontrado

components:
  schemas:
    ArrayOfProdutos:
      type: array
      items:
        $ref: '#/components/schemas/Produto'

    Produto:
      type: object
      properties:
        id:
          type: string
          example: "1"
        nome:
          type: string
          example: "Arroz"
        preco:
          type: number
          example: 12.5